% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/create_lin_data.R
\name{create_lin_data}
\alias{create_lin_data}
\title{Function to create data for linear model}
\usage{
create_lin_data(specs, data_set)
}
\arguments{
\item{specs}{A list with specifications that go into 'lp_lin' function.
lags_criterion: either NaN (given lag length) or 'BIC'|'AIC'}

\item{data_set}{A data frame with all endogenous variables.}
}
\value{
List with output from lm object.
}
\description{
Function to create data for linear model
}
