% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/create_nl_data.R
\name{create_nl_data}
\alias{create_nl_data}
\title{Function to create data for non-linear model.}
\usage{
create_nl_data(specs, data_set_df)
}
\arguments{
\item{specs}{A \link{list}(). Inputs are outlined in \link{lp_nl}.}

\item{data_set_df}{A data frame with all endogenous variables}
}
\value{
A list with lagged data for non-linear model.
}
\description{
Function to create data for non-linear model.
}
\author{
Philipp Adämmer
}
