% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/find_optim_lags.R
\name{find_optim_lags}
\alias{find_optim_lags}
\title{Title}
\usage{
find_optim_lags(yy, xx, specs)
}
\arguments{
\item{yy}{A matrix with 'endogenous' variable}

\item{xx}{A matrix with lagged 'exogenous' variables}

\item{specs}{}
}
\description{
Title
}
