% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_mat_chol.R
\name{get_mat_chol}
\alias{get_mat_chol}
\title{Compute structural shock matrix via Cholesky decomposition}
\usage{
get_mat_chol(y_lin, x_lin, endog_data, specs)
}
\arguments{
\item{y_lin}{A matrix with all endogenous variables.}

\item{x_lin}{A matrix with lagged endogenous variables.}

\item{endog_data}{A \link{data.frame} with all endogenous variables.}

\item{specs}{A list with specifications from \link{lp_lin} or \link{lp_nl}.}
}
\value{
Shock matrix (d)
}
\description{
Compute structural shock matrix via Cholesky decomposition with input variables
created in \link{lp_lin} or \link{lp_nl}.
}
\author{
Philipp Adämmer
}
\keyword{internal}
