% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_resids_ols.R
\name{get_resids_ols}
\alias{get_resids_ols}
\title{Compute residuals from OLS model}
\usage{
get_resids_ols(y_data, x_data)
}
\arguments{
\item{y_data}{A vector for the left hand variable.}

\item{x_data}{A matrix with right hand variables.}
}
\value{
A numeric vector with residuals from OLS regression.
}
\description{
Estimate residuals from OLS model.
}
\author{
Philipp Adämmer
}
\keyword{internal}
