% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_robust_cov_panel.R
\name{get_robust_cov_panel}
\alias{get_robust_cov_panel}
\title{Function to get robust covariance matrix for panel data}
\usage{
get_robust_cov_panel(panel_results, specs)
}
\arguments{
\item{panel_results}{Plm object from estimation}

\item{specs}{List with specifications}
}
\value{
Object with robust covariance matrix
}
\description{
Function to get robust covariance matrix for panel data
}
