% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_std_err_tsls.R
\name{get_std_err_tsls}
\alias{get_std_err_tsls}
\title{Compute standard errors for 2SLS}
\usage{
get_std_err_tsls(yy, xx, lag_nw, k, zz, specs)
}
\arguments{
\item{yy}{A matrix with all endogenous variables.}

\item{xx}{A matrix with lagged endogenous variables.}

\item{lag_nw}{Integer. Number of lags for Newey and West}

\item{k}{Integer. Value of column to choose from}

\item{specs}{A list with specifications from \link{lp_lin} or \link{lp_nl}.}
}
\value{
List. First element contains the estimated standard errors, the second element contains the OLS point estimates.
}
\description{
Compute 2SLS standard errors.#'
}
\author{
Philipp Adämmer
}
\keyword{internal}
