% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{get_vals_lagcrit}
\alias{get_vals_lagcrit}
\title{Compute values for lag length criterion}
\arguments{
\item{y}{List with left (endogenous) variables.}

\item{x}{List with right (exogenous) variables.}

\item{lag_crit}{Integer: 'AICc'= 1, 'AIC' = 2, 'BIC' = 3.}

\item{h}{Integer.}

\item{k}{Integer.}

\item{max_lags}{Integer.}
}
\description{
Compute values for 'AICc', 'AIC', or 'BIC'.
}
\keyword{internal}
