% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_var_lagcrit.R
\name{get_var_lagcrit}
\alias{get_var_lagcrit}
\title{Computes AICc, AIC and BIC for VAR}
\usage{
get_var_lagcrit(endog_data, specs = NULL)
}
\arguments{
\item{endog_data}{A \link{data.frame} with endogenous variables for the VAR}

\item{specs}{A \link{list} created in \link{lp_lin}}
}
\value{
A list with lag length criteria
}
\description{
Computes AICc, AIC and BIC for VAR models.
}
\references{
Akaike, H. (1974). "A new look at the statistical model identification",
\emph{IEEE Transactions on Automatic Control}, 19 (6): 716–723.

Hamilton, J. D. (1994). "Time Series Analysis."
Princeton: Princeton University Press.

Hurvich, C. M., and Tsai, C.-L. (1989), "Regression and time series model selection in small samples",
\emph{Biometrika}, 76(2): 297–307

Lütkepohl, H. (2005). "New Introduction to Multiple Time Series Analysis.",
New York: Springer.

Schwarz, Gideon E. (1978). "Estimating the dimension of a model",
\emph{Annals of Statistics}, 6 (2): 461–464.
}
\keyword{internal}
