% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hp_filter.R
\name{hp_filter}
\alias{hp_filter}
\title{Detrend a times series via the Hodrick-Prescott filter}
\usage{
hp_filter(x, lambda)
}
\arguments{
\item{x}{A something.}

\item{lambda}{A number.}
}
\value{
ret_results A list with results.
}
\description{
Estimates cyclicyal and trend component with HP-filter by Hodrick and Prescott (1997).
}
\references{
Hodrick, R.J., and Prescott, E. C. Prescott (1997) "Postwar U.S. Business Cycles: An Empirical Investigation."
 \emph{Journal of Money, Credit and Banking}, 29(1), 1-16.
}
