% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{hp_filter_c}
\alias{hp_filter_c}
\title{Detrend a times series via the Hodrick-Prescott filter}
\usage{
hp_filter_c(x, lambda)
}
\arguments{
\item{x}{A something}

\item{lambda}{A something
Hodrick, R.J., and Prescott, E. C. Prescott (1997) "Postwar U.S. Business Cycles: An Empirical Investigation."
\emph{Journal of Money, Credit and Banking}, 29(1), 1-16.}
}
\description{
Estimates cyclical and trend component with HP-filter by Hodrick and Prescott (1997).
The function is based on the R-code in the archived package "mFilter" by  Mehmet Balcilar.
}
