% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data-documentation.R
\docType{data}
\name{interest_rules_var_data}
\alias{interest_rules_var_data}
\title{A data frame containing data to estimate the effects of interest rate rules for monetary policy}
\format{A data frame with 193 quarterly observations (rows) and 3 variables (columns):
\describe{
  \item{GDP_gap}{Percentage difference between real GDP and potential GDP (Congressional Budget Office).}
  \item{Infl}{Inflation: Percentage change in the GDP, chain weighted price index at annual rate.}
  \item{FF}{Federal funds rate: quarterly average of daily rates.}
}
Sample: 1955:I - 2003:I}
\source{
\url{https://www.aeaweb.org/articles?id=10.1257/0002828053828518}
}
\usage{
interest_rules_var_data
}
\description{
A data frame containing data to estimate the effects of interest rate rules for monetary policy
}
\keyword{datasets}
