% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lm_function.R
\name{lm_function}
\alias{lm_function}
\title{Creates a data frame with lagged exogenous variables.}
\usage{
lm_function(m)
}
\arguments{
\item{data}{A data frame.}

\item{lags}{numeric: The number of lags.}
}
\value{
A data frame.
}
\description{
Creates a data frame with lagged exogenous variables.
}
