% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lp_lin.R
\name{lp_lin}
\alias{lp_lin}
\title{Title}
\usage{
lp_lin(data_set_estim, specs)
}
\arguments{
\item{data_set_estim}{A data.frame/tibble}

\item{specs}{List with stuff}
}
\value{
List with mean... SaetÄTER
}
\description{
Title
}
