% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lp_nl.R
\name{lp_nl}
\alias{lp_nl}
\title{lp_nl}
\usage{
lp_nl(data_set_df, specs)
}
\arguments{
\item{specs}{A list with specifications that go into 'lin_lp' or 'nl_lp' function.
specs$lags_criterion: Either NaN (given lag length) or 'AICc'|BICc'|'BIC'
specs$lags_lin:       Lag length
specs$trend:          1 (no trend), 2 (trend), 3 (quadratic trend)
specs$max_lags:       Maximum number of lags to use for lag length criteria
specs$lamdba:    Value of lambda for HP filter
specs$hp_filter: 1 = Use HP-filter to extract standardized time series
                 0 = Use pre-defined time series for transition function}

\item{data_set_estim}{A data.frame/tibble}
}
\value{
A list with irfs from local projections and newey west standard errors
}
\description{
lp_nl
}
