% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lp_nl_iv.R
\name{lp_nl_iv}
\alias{lp_nl_iv}
\title{Compute nonlinear impulse responses with instrument variable approach}
\usage{
lp_nl_iv(endog_data, lags_nl = NULL, instr = NULL, exog_data = NULL,
  lags_exog = NULL, contemp_data = NULL, lags_criterion = NULL,
  max_lags = NULL, trend = NULL, shock_type = NULL, confint = NULL,
  hor = NULL, switching = NULL, use_hp = NULL, lambda = NULL,
  gamma = NULL, num_cores = NULL)
}
\arguments{
\item{endog_data}{A \link{data.frame} containing all endogenous variables for the VAR. The column order
is used for the Cholesky decomposition.}

\item{lags_nl}{NaN or integer. Number of lags for nonlinear VAR (if \emph{lags_criterion} = NaN). NaN if lag length criterion is given.}

\item{instr}{One column \link{data.frame} including the instrument to shock with.
The row length has to be the same as \emph{endog_data}.}

\item{exog_data}{Null or a \link{data.frame}, containing exogenous data. The row length has to be the same as \emph{endog_data}.
The exogenous variables will not be separated into two regimes as the endogenous variables.}

\item{lags_exog}{Null or an integer, indicating the number of lags for exogenous data.}

\item{contemp_data}{A \link{data.frame}, containing exogenous data with contemporaneous impact. This data will not be lagged.
The row length has to be the same as \emph{endog_data}.}

\item{lags_criterion}{NaN or character. NaN means that the number of lags
will be given at \emph{lags_nl} and \emph{lags_lin}. The lag length criteria are 'AICc', 'AIC' and 'BIC'.}

\item{max_lags}{NaN or integer. Maximum number of lags (if \emph{lags_criterion} = 'AICc', 'AIC', 'BIC'). NaN otherwise.}

\item{trend}{Integer. Include no trend =  0 , include trend = 1, include trend and quadratic trend = 2.}

\item{shock_type}{Integer. Standard deviation shock = 0, unit shock = 1.}

\item{confint}{Double. Width of confidence bands. 68\% = 1; 90\% = 1.65; 95\% = 1.96.}

\item{hor}{Integer. Number of horizons for impulse responses.}

\item{switching}{Numeric vector. A column vector with the same length as \emph{endog_data}. This series can either
be decomposed via the Hodrick-Prescott filter (see Auerbach and Gorodnichenko, 2013) or
directly plugged into the following smooth transition function:
\deqn{ F_{z_t} = \frac{exp(-\gamma z_t)}{1 + exp(-\gamma z_t)} }
Warning: \eqn{F_{z_t}} will be lagged in \link{create_nl_data} by one and then multiplied with the data.
If the variable shall not be lagged, the vector has to be given with a lead of one.
The data for the two regimes are: \cr
Regime 1 = (1-\eqn{F(z_{t-1})})*y_{t-p}, \cr
Regime 2 = \eqn{F(z_{t-1})}*y_{t-p}.}

\item{use_hp}{Integer. No HP-filter = 0. Use HP-filter = 1.}

\item{lambda}{Double. Value of \eqn{\lambda} for the Hodrick-Prescott filter if HP-filter is applied.}

\item{gamma}{Double. Positive number which is used in the transition function.}

\item{num_cores}{Integer. The number of cores to use for the estimation. If no number is set, the function will
use the maximum number of available cores less one.}
}
\description{
Compute nonlinear impulse responses with local projections by Jordà (2005) and impulse variable approach as in
Ramey and Zubairy (2018). The can be separated into two states via a smooth transition function, applied in Auerbach and Gorodnichenko (2012).
}
\examples{
\donttest{

# This example replicates results from the Supplementary Appendix
# by Ramey and Zubairy (2018) (RZ-18). It evaluates results from Auerbach and
# Gorodnichenko (2012) (AG-12) with local projections by Jordá (2005).
# The data is taken from \\url{https://www.journals.uchicago.edu/doi/10.1086/696277}{JoPE}


# Load and prepare data
# The sample length of RZ-2018 is 1948:III-2008:III
 ag_data           <- ag_data
 sample_start      <- 7
 sample_end        <- dim(ag_data)[1]
 endog_data        <- ag_data[sample_start:sample_end, 3:5]

# The shock is created by RZ-18 and available as supplementary data
# to their paper
 shock             <- ag_data[sample_start:sample_end, 7]

# AG-12 also include four lags of the 7-quarter moving average growth rate
# as exogenous regressors in their model (see RZ-18)
 exog_data         <- ag_data[sample_start:sample_end, 6]

# Choose the 7-quarter moving average growth rate as switching variable.
# and adjust it to have suffiently long recession periods.
 switching_variable <- ag_data$GDP_MA[sample_start:sample_end] - 0.8

# Estimate local projections
 results_nl_iv <- lp_nl_iv(endog_data,
                           lags_nl           = 3,
                           instr             = shock,
                           exog_data         = exog_data,
                           lags_exog         = 4,
                           contemp_data      = NULL,
                           lags_criterion     = NaN,
                           max_lags          = NaN,
                           trend             = 0,
                           shock_type        = 1,
                           confint           = 1.96,
                           hor               = 20,
                           switching         = switching_variable,
                           use_hp            = 0,
                           lambda            = NaN, # Ravn and Uhlig (2002):
                                                    # Anuual data    = 6.25
                                                    # Quarterly data = 1600
                                                    # Monthly data   = 129,600
                           gamma             = 3)

# Make and save plots
 plots_nl_iv <- plot_nl(results_nl_iv)

# Show single impulse responses
# Compare with Figure 12 from Supplementary Appendix of RZ-18.
 plot(plots_nl_iv$gg_s1[[1]])
 plot(plots_nl_iv$gg_s2[[1]])

# Prepare to show all plots
 library(ggpubr)
 library(gridExtra)

 s1_plots <- sapply(plots_nl_iv$gg_s1, ggplotGrob)
 s2_plots <- sapply(plots_nl_iv$gg_s2, ggplotGrob)

# Show all responses of state 1
 marrangeGrob(s1_plots, nrow = ncol(endog_data), ncol = 1, top = NULL)

# Show all responses of state 2
 marrangeGrob(s2_plots, nrow = ncol(endog_data), ncol = 1, top = NULL)

}
}
\references{
Akaike, H. (1974). "A new look at the statistical model identification", \emph{IEEE Transactions on Automatic Control}, 19 (6): 716–723.

Auerbach, A. J., and  Gorodnichenko Y. (2012). "Measuring the Output Responses to Fiscal Policy."
\emph{American Economic Journal: Economic Policy}, 4 (2): 1-27.

Auerbach, A. J., and Gorodnichenko Y. (2013). "Fiscal Multipliers in Recession and Expansion."
\emph{NBER Working Paper Series}. Nr 17447.

Hurvich, C. M., and Tsai, C.-L. (1993) “A Corrected Akaike Information Criterion for
Vector Autoregressive Model Selection.” \emph{Journal of Time Series Analysis}, 1993, 14(3):
271–79.

Jordà, Ò. (2005) "Estimation and Inference of Impulse Responses by Local Projections."
\emph{American Economic Review}, 95 (1): 161-182.

Newey W.K., and West K.D. (1987). “A Simple, Positive-Definite, Heteroskedasticity and
Autocorrelation Consistent Covariance Matrix.” \emph{Econometrica}, 55, 703–708.

Ramey, V.A., Zubairy, S. (2018). "Government Spending Multipliers in Good Times
and in Bad: Evidence from US Historical Data." \emph{Journal of Political Economy},
126(2): 850 - 901.

Schwarz, Gideon E. (1978). "Estimating the dimension of a model", \emph{Annals of Statistics}, 6 (2): 461–464.
}
\seealso{
\url{https://adaemmerp.github.io/lpirfs/README_docs.html}
}
\author{
Philipp Adämmer
}
