% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lp_panel_lin.R
\name{lp_panel_lin}
\alias{lp_panel_lin}
\title{Estimate impulse responses with local projections for panel data}
\usage{
lp_panel_lin(data_set = NULL, sample = "Full", endog_data = NULL,
  cumul_mult = TRUE, shock = NULL, diff_shock = TRUE,
  iv_reg = FALSE, instrum = NULL, robust_se = FALSE,
  c_exog_data = NULL, l_exog_data = NULL, lags_exog_data = NaN,
  c_fd_exog_data = NULL, l_fd_exog_data = NULL,
  lags_fd_exog_data = NaN, confint = NULL, hor = NULL)
}
\arguments{
\item{data_set}{A \link{data.frame}, containing the panel data set. The first column has to be the
variable denoting the cross-section. The second column has to be the
variable denoting the time-section.}

\item{sample}{Boolean or character. Use full sample? TRUE (default). To estimate only a subset of the data, you have to provide
a sequence of the dates, years, etc. to use.}

\item{endog_data}{Character. Name of the endogenous variable. You can only provide one endogenous variable.}

\item{cumul_mult}{Boolean. Estimate cumulative multipliers? TRUE or FALSE.}

\item{shock}{The name of the variable to shock with.}

\item{diff_shock}{Boolean. Take first differences of the shock variable? TRUE or FALSE.}

\item{iv_reg}{Boolean. Use instrument variable approach? TRUE or FALSE}

\item{instrum}{NULL or Character. The name(s) of the instrument variable(s)}

\item{robust_se}{Boolean. Estimate robust standard errors á la Driscoll and Kraay (1998)? TRUE or FALSE.}

\item{c_exog_data}{NULL or Character. Names of exogenous variables with contemporaneous impact.}

\item{l_exog_data}{NULL or Character. Names of exogenous variables with lagged impact.}

\item{lags_exog_data}{Integer. Lag length for the exogenous data with lagged impact.}

\item{c_fd_exog_data}{NULL or Character. Names of exogenous variables with contemporaneous impact of first differences.}

\item{l_fd_exog_data}{NULL or Character. Names of exogenous variables with lagged impact of first differences.}

\item{lags_fd_exog_data}{NaN or Integer. Number of lags for first differenced variables.}

\item{confint}{Double. Width of confidence bands. 68\% = 1; 90\% = 1.65; 9\% = 1.96.}

\item{hor}{Integer. Number of horizons for impulse responses.}
}
\value{
A list containing:

\item{irf_lin_mean}{A \link{matrix}, containing the impulse responses.
                  The columns are the horizons.}

\item{irf_lin_low}{A \link{matrix}, containing all lower confidence bands of
                   the impulse responses, based on robust standard errors by Driscoll
                   and Kraay (1998). Properties are equal to \emph{irf_lin_mean}.}

\item{irf_lin_up}{A \link{matrix}, containing all upper confidence bands of
                   the impulse responses, based on robust standard errors by Driscoll
                   and Kraay (1998). Properties are equal to \emph{irf_lin_mean}.}

\item{reg_summaries}{Regression output for each horizon.}

\item{xy_data_sets}{Created panel data set with endogenous and exogenous variable for each horizon.}

\item{specs}{A list with properties of \emph{endog_data} for the plot function. It also contains
            the data sets used for the estimations (xy_data_sets) and the regression outputs (reg_summaries).}
}
\description{
This function estimates impulse responses with local projections for panel data.
             It works a bit different than the other 'lp_' functions. Instead of providing the endogenous
             and exogenous data sepearately as a data.frame, you to provide the entire panel data set as a data.frame.
             You can then simply have to give the names of the different variables to use.
}
\examples{
\donttest{

# This example aims to replicate the STATA code for panel estimation, provided on
#  Òscar Jordà's website

# Load libraries to download and reade excel file
 library(httr)
 library(readxl)
 library(dplyr)

# Retrieve the Jordà-Schularick-Taylor Macrohistory Database from the macrohistory lab in Bonn
 url_jst <-"http://www.macrohistory.net/JST/JSTdatasetR3.xlsx"
 GET(url_jst, write_disk(jst_link <- tempfile(fileext = ".xlsx")))
 jst_data <- read_excel(jst_link, 2L)

# Swap the first two columns so that 'country' is the first and 'year' the second column
 jst_data <- jst_data \%>\%
             dplyr::select(country, year, everything())

# Prepare variables. This is based on the 'data.do' file on Òscar Jordà's website
# Select data
  data_set <- jst_data \%>\%
               mutate(stir     = stir)                         \%>\%
               mutate(mortgdp  = 100*(tmort/gdp))              \%>\%
               mutate(hpreal   = hpnom/cpi)                    \%>\%
               group_by(country)                               \%>\%
               mutate(hpreal   = hpreal/hpreal[year==1990][1]) \%>\%
               mutate(lhpreal  = log(hpreal))                  \%>\%

               mutate(lhpy     = lhpreal - log(rgdppc))        \%>\%
               mutate(lhpy     = lhpy - lhpy[year == 1990][1]) \%>\%
               mutate(lhpreal  = 100*lhpreal)                  \%>\%
               mutate(lhpy     = 100*lhpy)                     \%>\%
               ungroup()                                       \%>\%

               mutate(lrgdp    = 100*log(rgdppc))              \%>\%
               mutate(lcpi     = 100*log(cpi)) 		            \%>\%
               mutate(lriy     = 100*log(iy*rgdppc))           \%>\%
               mutate(cay      = 100*(ca/gdp))                 \%>\%
               mutate(tnmort   = tloans - tmort)               \%>\%
               mutate(nmortgdp = 100*(tnmort/gdp))             \%>\%

             # Prepare instrument
               mutate(lnarrow  = log(narrowm))                 \%>\%
               mutate(rlnarrow = lnarrow - lcpi)               \%>\%
               dplyr::select(country, year, mortgdp, stir, ltrate, lhpy,
                             lrgdp, lcpi, lriy, cay, nmortgdp, rlnarrow)


# Use sample from 1870 - 2013 AND exclude WWI and WWII
  sample <-   seq(1870, 2016)[which(!(seq(1870, 2016) \%in\%
                              c(seq(1914, 1918), seq(1939, 1947), seq(2014, 2016))))]

# Estimate panel model
results_panel <-  lp_panel_lin(data_set          = data_set,
                               sample            = sample,
                               endog_data        = "mortgdp",
                               cumul_mult        = TRUE,

                               shock             = "stir",
                               diff_shock        = TRUE,
                               iv_reg            = FALSE,
                               instrum           = NULL,
                               robust_se         = TRUE,

                               exog_data         = colnames(data_set)[3:11],
                               c_exog_data       = "cay",
                               l_exog_data       = NULL,
                               lags_exog_data    = NULL,
                               c_fd_exog_data    = colnames(data_set)[c(seq(4,9),11)],
                               l_fd_exog_data    = colnames(data_set)[c(seq(3,9),11)],
                               lags_fd_exog_data = 2,

                               confint           = 1.67,
                               hor               = 5)



# Create and add instrument to data_set

 instrument <- 0.9*data_set$stir + rnorm(length(data_set$stir), 0, sd(na.omit(data_set$stir)))
 data_set   <- data_set \%>\%
               mutate(instrument = instrument)


# Estimate panel model with iv
results_panel <-  lp_panel_lin(data_set          = data_set,
                               sample            = sample,
                               endog_data        = "mortgdp",
                               cumul_mult        = TRUE,

                               shock             = "stir",
                               diff_shock        = TRUE,
                               iv_reg            = TRUE,
                               instrum           = instrument,
                               robust_se         = TRUE,

                               exog_data         = colnames(data_set)[3:11],
                               c_exog_data       = "cay",
                               l_exog_data       = NULL,
                               lags_exog_data    = NULL,
                               c_fd_exog_data    = colnames(data_set)[c(seq(4,9),11)],
                               l_fd_exog_data    = colnames(data_set)[c(seq(3,9),11)],
                               lags_fd_exog_data = 2,

                               confint           = 1.67,
                               hor               = 5)

}

}
\references{
Driscoll, J.C., and Kraay, A.C. (1998). Consistent Covariance Matrix Estimation with Spatially Dependent
Panel Data, \emph{Review of Economics and Statistics}, 80(4), pp. 549–560.

Jordà, Ò. (2005). "Estimation and Inference of Impulse Responses by Local Projections."
\emph{American Economic Review}, 95 (1): 161-182.
}
\author{
Philipp Adämmer
}
