% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/package.R
\docType{package}
\name{lpirfs-package}
\alias{lpirfs-package}
\title{Local Projection Impulse response Functions}
\description{
Contains functions to estimate local projections as proposed by Jordà (2005)
}
\author{
Philipp Adämmer
}
