% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data-documentation.R
\docType{data}
\name{monetary_var_data}
\alias{monetary_var_data}
\alias{monetary_var_data}
\title{Data to estimate the effects of interest rate rules for monetary policy, applied in the AER article by Jordà (2005).
<doi:10.1257/0002828053828518>
Quarterly data to estimate the effects of interest rate rules for monetary policy.
Sample spans from 1955:I - 2003:I}
\format{A data frame with 193 observations (rows) and 3 variables:
\describe{
  \item{GDP_gap}{Percentage difference between real GDP and potential GDP (Congressional Budget Office.}
  \item{Infl}{Inflation: Percentage change in the GDP, chain weighted price index at annual rate.}
  \item{FF}{Federal funds rate: quarterly average of daily rates}
  ...
}}
\source{
\url{https://www.aeaweb.org/articles?id=10.1257/0002828053828518}

\url{https://www.aeaweb.org/articles?id=10.1257/0002828053828518}
}
\usage{
monetary_var_data

monetary_var_data
}
\description{
Data to estimate the effects of interest rate rules for monetary policy, applied in the AER article by Jordà (2005).
<doi:10.1257/0002828053828518>
Quarterly data to estimate the effects of interest rate rules for monetary policy.
Sample spans from 1955:I - 2003:I

Data for for the Christiano, Eichenbaum and Evans (1996) VAR,
applied in the AER article by Jordà (2005).
<doi:10.1257/0002828053828518>
Monthly data for a standard monetary VAR. Sample spans from 1960:01 - 2001:02.
}
\keyword{datasets}
