% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{newey_west}
\alias{newey_west}
\title{Compute OLS parameters and robust standard errors based on Newey-West estimator}
\usage{
newey_west(y, x, h)
}
\arguments{
\item{y}{Numeric vector.}

\item{x}{Numeric matrix.}

\item{h}{Integer.}
}
\value{
A list. The first element contains the estimated OLS parameters, the second element
the Newey West covariance matrix, the third element the estimated functions, the fourth element
the unscaled covariance matrix and the last element the meat estimator.
}
\description{
Compute OLS parameters and robust standard errors based on Newey and West (1987).
The function is based on the Matlab code by James P. LeSage.
}
\references{
Newey, W.K., and West, K.D. (1987). “A Simple, Positive-Definite, Heteroskedasticity and
Autocorrelation Consistent Covariance Matrix.” \emph{Econometrica}, 55, 703–708.
}
\keyword{internal}
