% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{ols_diagnost}
\alias{ols_diagnost}
\title{Compute diagnostics for OLS models}
\usage{
ols_diagnost(y, x)
}
\arguments{
\item{y}{Numeric vector.}

\item{x}{Numeric matrix.}
}
\value{
A list. The first element contains...
}
\description{
Compute OLS diagnostics such as R^2, adjusted R^2, AIC, etc.
}
\references{
Akaike, H. (1974). "A new look at the statistical model identification", \emph{IEEE Transactions on Automatic Control}, 19 (6): 716–723.
Hurvich, C. M., and Tsai, C.-L. (1989). "Regression and time series model selection
in small samples", Biometrika, 76(2): 297–307,

Schwarz, Gideon E. (1978). "Estimating the dimension of a model", \emph{Annals of Statistics}, 6 (2): 461–464.
}
\keyword{internal}
