% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lpirfs_lin_obj-methods-base.R
\name{plot.lpirfs_lin_panel_obj}
\alias{plot.lpirfs_lin_panel_obj}
\title{Base print() function to plot all impulse responses from linear lpirfs object}
\usage{
\method{plot}{lpirfs_lin_panel_obj}(x, ...)
}
\arguments{
\item{x}{An object of type 'lpirfs_lin_panel_obj'}

\item{...}{Additional arguments to be consistent with S3 print() function}
}
\description{
Base print() function to plot all impulse responses from linear lpirfs object
}
