% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_lin.R
\name{plot_lin}
\alias{plot_lin}
\title{Compute and display plots of linear impulse responses}
\usage{
plot_lin(results_lin)
}
\arguments{
\item{results_lin}{A \link{list} created with \link{lp_lin}() or \link{lp_lin_iv}().}
}
\value{
A list with (gg-)plots for linear impulse responses.
}
\description{
Compute and display linear impulse responses, estimated with \link{lp_lin}() and \link{lp_lin_iv}().
}
\examples{
\donttest{

# See examples for lp_lin() and lp_lin_iv().
 }
}
\author{
Philipp Adämmer
}
