% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_lin_irfs.R
\name{plot_lin_irfs}
\alias{plot_lin_irfs}
\title{Function to estimate linear irfs, estimated with lin_lp function}
\usage{
plot_lin_irfs(results_lin)
}
\arguments{
\item{results_lin}{}
}
\description{
Function to estimate linear irfs, estimated with lin_lp function
}
