% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_nl.R
\name{plot_nl}
\alias{plot_nl}
\title{Compute and display plots of nonlinear impulse responses}
\usage{
plot_nl(results_nl)
}
\arguments{
\item{results_nl}{A \link{list} created with \link{lp_nl}() or \link{lp_nl_iv}().}
}
\value{
A list with (gg-)plots for nonlinear impulse responses.
}
\description{
Compute and display (nonlinear) impulse responses, estimated with \link{lp_nl}() and \link{lp_nl_iv}().
}
\examples{
\donttest{
# Load package

 # See examples for lp_nl() and lp_nl_iv().

 }
}
\author{
Philipp Adämmer
}
