% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_nl_irfs.R
\name{plot_nl_irfs}
\alias{plot_nl_irfs}
\title{Compute (non-linear) irf plots}
\usage{
plot_nl_irfs(results_nl)
}
\arguments{
\item{results_nl}{A list with 3D arrays estimated in \link{lp_nl}.}
}
\value{
A list with plots for non-linear impulse responses
}
\description{
Compute (non-linear) irf plots irf plots estimated with \link{lp_nl}.
}
\author{
Philipp Adämmer
}
