% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/reduced_var.R
\name{reduced_var}
\alias{reduced_var}
\title{Estimate structural shock matrix via Cholesky decomposition.}
\usage{
reduced_var(y_lin, x_lin, data_set_df, specs)
}
\arguments{
\item{y_lin}{A data frame for the exogenous variables.}

\item{x_lin}{A list with exogenous variables.}

\item{data_set_df}{Data frame with all endogenous variables.}

\item{specs}{A list with specifications from \link{lp_lin}.}
}
\value{
Shock matrix (d)
}
\description{
Estimate structural shock matrix via Cholesky decomposition.
}
\author{
Philipp Adämmer
}
