% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/se_hc_panel_cluster.R
\name{se_hc_panel_cluster}
\alias{se_hc_panel_cluster}
\title{Create and returns a function to estimate hc and clustered standard errors for panel data.}
\usage{
se_hc_panel_cluster(func_name)
}
\arguments{
\item{func_name}{A character.}
}
\value{
A function.
}
\description{
Create and returns a function to estimate hc and clustered standard errors for panel data.
}
\references{
Croissant, Y., Millo, G. (2017). "Robust Standard Error Estimators for Panel Models: A Unifying Approach." \emph{Journal of Statistical Software}, 27(2), 1-43. doi:
	10.18637/jss.v082.i03 (URL: \url{https://www.jstatsoft.org/article/view/v082i03}).
}
\keyword{internal}
