% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/slp_conf.R
\name{slp_conf}
\alias{slp_conf}
\title{Estimates confidence bands by Newey West estimator}
\source{
\url{https://github.com/ctbrownlees/R-Package-lproj}
}
\usage{
slp_conf(obj, specs, l = 1)
}
\arguments{
\item{obj}{A list with values built in slp_estim.}

\item{specs}{A list with values, built in slp_lin.}

\item{l}{Integer. Numeric value.}
}
\value{
A \link{list}, named obj.
}
\description{
Compute confidence for smooth local projections by Newey and West (1987).
The function is based on the R-code by Barnichon and Brownless (2018), available on GitHub.
}
\references{
Barnichon, R., Brwonless, C. (2018), "Impulse Response Estimation By Smooth Local Projections",
emph{The Review of Economics and Statistics}, Forthcoming.

Newey, W.K., and West, K.D. (1987). “A Simple, Positive-Definite, Heteroskedasticity and
Autocorrelation Consistent Covariance Matrix.” \emph{Econometrica}, 55, 703–708.
}
\keyword{internal}
