% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/slp_estim.R
\name{slp_estim}
\alias{slp_estim}
\title{Function to compute linear smooth local projections}
\source{
\url{https://github.com/ctbrownlees/R-Package-lproj}
}
\usage{
slp_estim(y, w, x, hor, h1, r, K, type, lambda, specs)
}
\arguments{
\item{y}{A \link{matrix} with the endogenous variable}

\item{w}{A matrix with the endogenous variables (including constant, trend, etc.).}

\item{x}{A matrix with the variable to shock with.}

\item{hor}{Integer. Number of horizons for impulse responses.}

\item{h1}{Integer. Start of local projections. Default is 1.}

\item{r}{Integer. Order of the limit polynomial.}

\item{K}{Integer. Number for k-fold cross validation. The default value is 5.}

\item{type}{Character. Type of estimation. Default is "smooth" for smooth local projections. "reg" estimates regular impulse responses.}

\item{specs}{List. A list with values, built in slp_lin.}
}
\value{
A list with estimation results
}
\description{
Function to compute linear impulse responses with smooth local projection approach.
The function is based on the R-code by Barnichon and Brownless (2018), available on GitHub.
}
\references{
Barnichon, R., Brwonless, C. (2018), "Impulse Response Estimation By Smooth Local Projections",
\emph{The Review of Economics and Statistics}, Forthcoming.
}
\keyword{internal}
