% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/slp_lin.R
\name{slp_lin}
\alias{slp_lin}
\title{Compute linear smooth local projections}
\usage{
slp_lin(endog_data, shock = NULL, use_twosls = FALSE,
  lags_endog_lin = NULL, exog_data = NULL, lags_exog = NULL,
  contemp_data = NULL, trend = NULL, confint = NULL, hor = NULL,
  h1 = NULL, bdeg = NULL, r = NULL, shock_type = NULL, K = 5,
  use_cv = TRUE, lambda = 0, type = "smooth")
}
\arguments{
\item{endog_data}{A \link{data.frame}, containing the values of the dependent variable. This function only works for one endogenous
variable at a time}

\item{shock}{A one column \link{data.frame}, including the variable to shock with. The row length has to be the same as \emph{endog_data}.}

\item{use_twosls}{Boolean. Use two stage least squares? TRUE or FALSE.}

\item{lags_endog_lin}{Integer. Integer for number of lags for \emph{endog_data}.}

\item{exog_data}{A \link{data.frame}, containing exogenous variables. The row length has to be the same as \emph{endog_data}.}

\item{lags_exog}{NULL or Integer. Integer for the number of lags for the exogenous data.}

\item{contemp_data}{A \link{data.frame}, containing exogenous data with contemporaneous impact.
The row length has to be the same as \emph{endog_data}.}

\item{trend}{Integer. No trend =  0 , include trend = 1, include trend and quadratic trend = 2.}

\item{confint}{Double. Width of confidence bands. 68\% = 1; 90\% = 1.65; 95\% = 1.96.}

\item{hor}{Integer. Number of horizons for impulse responses.}

\item{h1}{Integer. Start of local projections. Default is 1.}

\item{bdeg}{Integer. Start values of knots to evaluate the design matrix for the B-splines. The default is 3.}

\item{r}{Integer. Order of the limit polynomial.}

\item{shock_type}{Integer. Standard deviation shock = 0, unit shock = 1.}

\item{K}{Integer. Number for k-fold cross validation. The default value is 5.}

\item{use_cv}{Boolean. Use cross validation?}

\item{lambda}{Numeric. A numeric number or numeric vector (if use_cv = TRUE) for shrinkage value(s).}

\item{type}{Character. Type of estimation. Default is "smooth" for smooth local projections. "reg" estimates regular impulse responses.}
}
\value{
A list with all values.
}
\description{
Compute linear impulse responses with smooth local projection approach by Barnichon and Brownless (2018).
}
\examples{

\donttest{

# Load package
library(lpirfs)

# Load data
data(us_macro)

# Make vector of possible shrinkage parameters
  lambda         <- c(0.0001,0.01,0.1,0.2,0.3,0.4,0.5,0.6,0.7,0.8,1.0,9,10)/1000


# Estimate irf with smooth local projections
smooth_results<- slp_lin(endog_data      = tibble(yg = us_macro$yg),
                          shock          = tibble(ir = us_macro$ir),
                          use_twosls     = FALSE,
                          lags_endog_lin = 4,
                          exog_data      = tibble(pi = us_macro$pi,ir = us_macro$ir),
                          lags_exog      = 4,
                          contemp_data   = tibble(yg = us_macro$yg, pi = us_macro$pi),
                          trend          = 0,
                          confint        = 1.28,
                          hor            = 20,
                          bdeg           = 3,
                          r              = 2,
                          K              = 5,
                          use_cv         = TRUE,
                          lambda         = lambda)

 # Make plots
 slp_lin_plots <- plot_lin(smooth_results)

 # This figure replicates the upper right plot in Figure 3 of
 # Barnichon and Brownless (2018)
 plot(slp_lin_plots[[1]])

}


}
\references{
Barnichon, R., Brwonless, C. (2018), "Impulse Response Estimation By Smooth Local Projections",
\emph{The Review of Economics and Statistics}, Forthcoming.
}
