% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lpirfs_obj-methods-base.R
\name{summary.lpirfs_lin_obj}
\alias{summary.lpirfs_lin_obj}
\title{Summary for linear lpirfs object}
\usage{
\method{summary}{lpirfs_lin_obj}(object, ...)
}
\arguments{
\item{object}{An object of type 'lpirfs_lin_obj'}

\item{...}{Additional arguments to be consistent with S3 print() function}
}
\description{
Summary for linear lpirfs object
}
