% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/switching_series.R
\name{switching_series}
\alias{switching_series}
\title{Function to estimate switching series/variables. It uses a smooth transition function as
Auerbach & Gorodnichenko (2012)  <doi:10.1257/pol.4.2.1>.
The standardized time series is either pre-defined or will be estimated via
the Hodrick-Prescott as suggested by Ramey & Zubairy (2018) <doi:10.1086/696277>}
\usage{
switching_series(switching_data, specs)
}
\arguments{
\item{switching_data}{A vector with normalized time series to construct switching variable}

\item{specs}{A list with specifications that go into 'lin_lp' or 'nl_lp' function.
specs$lamdba:    Value of lambda for HP filter
specs$hp_filter: 1 = Use HP-filter to extract standardized time series
                 0 = Use pre-defined time series for transition function}
}
\value{
fz: Vector with values from smooth transition function
}
\description{
Function to estimate switching series/variables. It uses a smooth transition function as
Auerbach & Gorodnichenko (2012)  <doi:10.1257/pol.4.2.1>.
The standardized time series is either pre-defined or will be estimated via
the Hodrick-Prescott as suggested by Ramey & Zubairy (2018) <doi:10.1086/696277>
}
\author{
Philipp Adämmer
}
