% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data-documentation.R
\docType{data}
\name{us_macro}
\alias{us_macro}
\title{Data to estimate the effects of a monetary shock}
\format{A \link{tibble} with 494 monthly observations (rows) and 6 variables (columns):
\describe{
  \item{yg}{GDP growth.}
  \item{pi}{Inflation rate.}
  \item{ir}{Interest rate.}
  }
Sample: 1959:01 - 2007:02.}
\source{
\url{https://github.com/ctbrownlees/R-Package-lproj}
}
\usage{
us_macro
}
\description{
A data.frame, containing quarterly data to estimate the effects of a monetary shock.
The data is used by Barnichon and Brownless (2018)
}
\references{
Barnichon, R., Brwonless, C. (2018), "Impulse Response Estimation By Smooth Local Projections",
\emph{The Review of Economics and Statistics}, Forthcoming.
}
\keyword{datasets}
