% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{var_one}
\alias{var_one}
\title{Compute VAR to prewhite estimating functions for Newey West estimator.}
\usage{
var_one(VAR_Data)
}
\arguments{
\item{VAR_Data}{Matrix.}
}
\value{
A list. The first element contains the slope parameters of the VAR(1), the sedond element contains
the residuals and the third element the inverted slope parameter matrix.
}
\description{
Compute Newey-West estimator with prewhitened estimation functions.
}
\references{
Andrews, D.W. and Monahan, J.C. (1992). An improved heteroskedasticity and
autocorrelation consistent covariance matrix estimator. \emph{Econometrica}, pp.953-966.

Newey, W.K., and West, K.D. (1987). “A Simple, Positive-Definite, Heteroskedasticity and
Autocorrelation Consistent Covariance Matrix.” \emph{Econometrica}, 55, 703–708.
}
\keyword{internal}
