% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/create_lags.R
\name{create_lags}
\alias{create_lags}
\title{Creates a data frame with lagged exogenous variables.}
\usage{
create_lags(data, lags)
}
\arguments{
\item{data}{A data frame with variables in columns.}

\item{lags}{Numeric value for the number of lags.}
}
\value{
lag_data: A data frame with lagged values from 'data'.
}
\description{
Creates a data frame with lagged exogenous variables.
}
