% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/create_lags.R
\name{create_lags}
\alias{create_lags}
\title{Compute a data frame with lagged exogenous variables}
\usage{
create_lags(data, lags)
}
\arguments{
\item{data}{A \link{data.frame}().}

\item{lags}{Integer for the number of lags.}
}
\value{
Returns a \link{data.frame} with lagged values.
}
\description{
Create a \link{data.frame}() with lagged exogenous variables.
}
\author{
Philipp Adämmer
}
\keyword{internal}
