% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/create_lin_data.R
\name{create_lin_data}
\alias{create_lin_data}
\title{Function to create data for linear model.}
\usage{
create_lin_data(specs, data_set_df)
}
\arguments{
\item{specs}{A list with specifications that go into 'lp_lin' or 'lp_nl' function.
specs$lags_criterion: Either NaN (given lag length) or 'AICc'|'AIC'|'BIC'|
specs$lags_lin:       Lag length
specs$trend:          1 (no trend), 2 (trend), 3 (quadratic trend)
specs$max_lags:       Maximum number of lags to determine lag length criteria}

\item{data_set_df}{A data frame with all endogenous variables for VAR.}
}
\value{
List with filled matrices of lagged left (y_lin) and right hand side (x_lin) variables
}
\description{
Function to create data for linear model.
}
\author{
Philipp Adämmer
}
