% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/create_lin_data.R
\name{create_lin_data}
\alias{create_lin_data}
\title{Compute data for linear model}
\usage{
create_lin_data(specs, endog_data)
}
\arguments{
\item{specs}{A \link{list}(). Inputs are created in \link{lp_lin_iv}.}

\item{endog_data}{A \link{data.frame} with dependent variables.}
}
\value{
List with filled matrices of lagged left (y_lin) and right hand side (x_lin) variables.
}
\description{
Function to create data for linear model.
}
\author{
Philipp Adämmer
}
\keyword{internal}
