% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/create_nl_data.R
\name{create_nl_data}
\alias{create_nl_data}
\title{Function to create data for non-linear model.}
\usage{
create_nl_data(specs, data_set_estim)
}
\arguments{
\item{specs}{A list with specifications that go into 'lin_lp' or 'nl_lp' function.
specs$lags_criterion: Either NaN (given lag length) or 'BIC'|'AIC'
specs$lags_lin:       Lag length
specs$trend:          1 (no trend), 2 (trend), 3 (quadratic trend)
specs$max_lags:       Maximum number of lags to use for lag length criteria}

\item{data_set_estim}{}
}
\value{
A list with lagged data for non-linear model.
}
\description{
Function to create data for non-linear model.
}
\author{
Philipp Adämmer
}
