% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/create_nl_data.R
\name{create_nl_data}
\alias{create_nl_data}
\title{Compute data for nonlinear model with instrument variable approach}
\usage{
create_nl_data(specs, endog_data)
}
\arguments{
\item{specs}{A \link{list}(). Inputs are created in \link{lp_nl_iv}.}

\item{endog_data}{A \link{data.frame} with dependent variables.}
}
\value{
List with filled matrices of lagged left (y_nl) and right hand side (x_nl) variables.
}
\description{
Function to create data for nonlinear model with instrument variable approach.
}
\author{
Philipp Adämmer
}
\keyword{internal}
