% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{find_lag_c}
\alias{find_lag_c}
\title{Determine optimal lag length}
\arguments{
\item{y}{List with left (endogenous) variables}

\item{x}{List with right (exogenous) variables}

\item{lag_crit}{Integer: 'AICc'= 1, 'AIC' = 2, 'BIC' = 3.}
}
\description{
Determine optimal lag length based on 'AICc', 'AIC', or 'BIC' criterion.
}
