% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{get_vals_lagcrit}
\alias{get_vals_lagcrit}
\title{Compute values for lag length criteria}
\usage{
get_vals_lagcrit(y, x, lag_crit, h, k, max_lags, n_obs)
}
\arguments{
\item{y}{List with left (endogenous) variables.}

\item{x}{List with right (exogenous) variables.}

\item{lag_crit}{Integer: 'AICc'= 1, 'AIC' = 2, 'BIC' = 3.}

\item{h}{Integer.}

\item{k}{Integer.}

\item{max_lags}{Integer.}
}
\description{
Compute values for 'AICc', 'AIC', or 'BIC'.

Schwarz, Gideon E. (1978). "Estimating the dimension of a model", \emph{Annals of Statistics}, 6 (2): 461–464.
}
\references{
Akaike, H. (1974). "A new look at the statistical model identification", \emph{IEEE Transactions on Automatic Control}, 19 (6): 716–723.
}
\keyword{internal}
