% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{hp_filter}
\alias{hp_filter}
\title{Decompose a times series via the Hodrick-Prescott filter}
\usage{
hp_filter(x, lambda)
}
\arguments{
\item{x}{Column matrix with numeric values.}

\item{lambda}{Numeric value.}
}
\value{
A list. The first element contains the cyclical component and the second element the trend component.
}
\description{
Estimates cyclical and trend component with HP-filter by Hodrick and Prescott (1997).
The function is based on the function \emph{hpfilter} from the package \emph{mFilter} by Mehmet Balcilar.
}
\references{
Hodrick, R.J., and Prescott, E. C. Prescott (1997) "Postwar U.S. Business Cycles: An Empirical Investigation."
\emph{Journal of Money, Credit and Banking}, 29(1), 1-16.
}
