% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lm_function.R
\name{lm_function}
\alias{lm_function}
\title{OLS function}
\usage{
lm_function(y_vector, x_data)
}
\arguments{
\item{y_vector}{A vector for the left hand variable.}

\item{x_lin}{A data frame with exogenous variables.}
}
\value{
List with output from lm object.
}
\description{
OLS function
}
