% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lm_function.R
\name{lm_function}
\alias{lm_function}
\title{lm function which returns OLS parameter estimates}
\usage{
lm_function(y_data, x_data)
}
\arguments{
\item{y_data}{A vector for the left hand variable.}

\item{x_data}{A data frame with exogenous variables.}
}
\value{
List with output from lm object.
}
\description{
lm function which returns OLS parameter estimates
}
