% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lm_function.R
\name{lm_function}
\alias{lm_function}
\title{Estimate residuals from OLS model.}
\usage{
lm_function(y_data, x_data)
}
\arguments{
\item{y_data}{A vector for the left hand variable.}

\item{x_data}{A data frame with exogenous variables.}
}
\value{
Vector with residuals from OLS regression.
}
\description{
Estimate residuals from OLS model.
}
\author{
Philipp Adämmer
}
