% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lm_function.R
\name{lm_function}
\alias{lm_function}
\title{Function to estimate VAR by OLS: equation for equation.}
\usage{
lm_function(y_data, x_lin, specs)
}
\arguments{
\item{y_data}{A data frame for the exogenous variables.}

\item{x_lin}{A data frame with exogenous variables.}

\item{specs}{List with general variables for estimation.}
}
\value{
List with output from lm object.
}
\description{
Function to estimate VAR by OLS: equation for equation.
}
