% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lp_lin.R
\name{lp_lin}
\alias{lp_lin}
\title{Compute (linear) impulse responses}
\usage{
lp_lin(data_set_df, specs)
}
\arguments{
\item{data_set_df}{A \link{data.frame} containing all endogenous variabls for the VAR.}

\item{specs}{A \link{list} with the following inputs:

\itemize{
\item{lags_criterion: NaN (Lag length is given), 'AICc', 'AIC' or 'BIC'.}
\item{lags_lin: Number of (linear) VAR lags (if \emph{lags_criterion} = NaN).}
\item{max_lags: Maximum number of lags if lags_criterion is given.}
\item{trend: No trend =  0 , Include trend = 1, Include trend and quadratic trend = 2.}
\item{shock_type: Standard deviation shock = 0, Unit shock = 1.}
\item{confint: Width of confidence bands. 68\% = 1, 90\% = 1.65, 95\% = 1.96.}
\item{hor: Horizons for irfs.}
}}
}
\value{
A list with impulse responses and their corresponding Newey West (1987) standard errors.
It also returns an updated list with the data frames specifications for the plot function.

\item{irf_lin_mean:}{A three 3D \link{array}, containing all impulse responses for all endogenous variables.
                    The last array dimension denotes the variable which shocks. The row in each corresponding matrix
                               denotes the respones of the \emph{ith} variable as ordered in data_set_df. The matrices columns denote the horizons.
                               For example, lp_lin$irf_lin_mean[, , 1] returns a KXH matrix, where K is the number of endogenous variables
                               and H the number of horizons. '1' denotes the variable which shocks.}

\item{irf_lin_low:}{A three 3D \link{array}, containing all lower confidence bands based on Newey West (1987).
                               Properties are equal to irf_lin_mean.}

\item{irf_lin_up:}{A three 3D \link{array}, containing all upper confidence bands based on Newey West (1987).
                               Properties are equal to irf_lin_mean.}

\item{specs:}{An updated list of \emph{specs} with updated entries for the plot function.}
}
\description{
Compute impulse responses with local projections by Jordà (2005).
}
\examples{
# Create list for function input
  specs <- list()

# Specify inputs
  specs$lags_lin       <- 12
  specs$lags_criterion <- NaN
  specs$max_lags       <- 2
  specs$trend          <- 1
  specs$shock_type     <- 1
  specs$confint        <- 1.96
  specs$hor            <- 24

# Estimate model and save results
 results_lin <- lpirfs::lp_lin(data_set_df, specs)
}
\references{
Jordà, O. (2005) "Estimation and Inference of Impulse Responses by Local Projections."
\emph{American Economic Review}, 95 (1): 161-182.

Newey W.K., West K.D. (1987). “A Simple, Positive-Definite, Heteroskedasticity and
Autocorrelation Consistent Covariance Matrix.” Econometrica, 55, 703–708.
}
