% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lp_nl.R
\name{lp_nl}
\alias{lp_nl}
\title{Compute (non-linear) impulse responses}
\usage{
lp_nl(data_set_df, specs)
}
\arguments{
\item{data_set_df}{A \link{data.frame}() containing all endogenous variables for the VAR. The column order
is also used for the Cholesky decomposition.}

\item{specs}{A \link{list}() with the following inputs:

\itemize{
\item{\strong{lags_criterion} NaN or character ('AICc', 'AIC' or 'BIC'). NaN means that the number of lags
        has to be given at \emph{lags_nl}. The character denotes the lag length criterion.}
\item{\strong{lags_nl} Integer. Number of lags for (non-linear) VAR (if \emph{lags_criterion} = NaN).}
\item{\strong{max_lags} Integer. Maximum number of lags (if \emph{lags_criterion} = 'AICc', 'AIC', 'BIC').}
\item{\strong{trend} Integer. Include no trend =  0 , include trend = 1, include trend and quadratic trend = 2.}
\item{\strong{shock_type} Integer. Standard deviation shock = 0, Unit shock = 1.}
\item{\strong{confint} Double. Width of confidence bands. 68\% = 1; 90\% = 1.65; 95\% = 1.96.}
\item{\strong{hor} Integer. Number of horizons for impulse responses. }
\item{\strong{switching} Vector. A column vector with the same length as \emph{data_set_df}. This series (\eqn{z_t}) can either
              be first decomposed by the Hodrick-Prescott filter as proposed in Ramey and Zubairy (2018) or
              directly plugged into the smooth transition function:
              \deqn{ F_{z_t}) = \frac{exp(-\gamma z_t)}{1 + exp(-\gamma z_t)} }
              Warning: \eqn{F_{z_t}} will first be lagged in \link{create_nl_data} by one and then multiplied with the data.
              If the variable shall not be lagged, it has to be given with a lead of one.
              The transition states are defined as: \cr
              State 1 = 1-\eqn{F(z_t)}, state 2 = \eqn{F(z_t)}
              Please read Auerbach and Gorodnichenko (2012) for further details.
                }
\item{\strong{hp_filter} Integer. No HP-filter = 0. Use HP-filter = 1. }
\item{\strong{lambda} Double. Value of \eqn{\lambda} for the Hodrick-Prescott filter if \emph{hp_filter} = 1. }
\item{\strong{gamma} Double. Value of \eqn{\gamma} which is used in the transition function in \link{switching_variable}. }
}}
}
\description{
Compute non-linear impulse responses with local projections by Jordà (2005). The
data are separated into two states with a smooth transition function as in by Auerbach and Gorodnichenko (2012).
}
\examples{
data("monetary_var_data")

# Create list for input
specs <- list()
# Fill list
specs$lags_nl        <- 4
specs$lags_criterion <- NaN
specs$max_lags       <- 2
specs$trend          <- 1
specs$shock_type     <- 1

# Specifications for switching variable
specs$switching      <- data_set_df$EM
specs$hp_filter      <- 1
specs$lambda         <- 1600
specs$gamma          <- -3
# Horizons and cinfidence intervals
specs$confint        <- 1.96
specs$hor            <- 24
# Estimate model and save results
 results_lin <- lpirfs::lp_lin(data_set_df, specs)
}
\references{
Auerbach, A. J., and  Gorodnichenko Y. (2012). "Measuring the Output Responses to Fiscal Policy."
\emph{American Economic Journal: Economic Policy}, 4 (2): 1-27.

Jordà, O. (2005) "Estimation and Inference of Impulse Responses by Local Projections."
\emph{American Economic Review}, 95 (1): 161-182.

Newey W.K., and West K.D. (1987). “A Simple, Positive-Definite, Heteroskedasticity and
Autocorrelation Consistent Covariance Matrix.” \emph{Econometrica}, 55, 703–708.

Ramey, V.A., and Zubairy, S. (2018). "Government Spending Multipliers in Good Times and in Bad:
Evidence from US Historical Data." \emph{Journal of Political Economy}, 126 (2), 850-901.
}
