% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lp_panel_lin.R
\name{lp_panel_lin}
\alias{lp_panel_lin}
\title{Estimate impulse responses with local projections for panel data}
\usage{
lp_panel_lin(data_set = NULL, sample = "Full", endog_data = NULL,
  cumul_mult = TRUE, shock = NULL, diff_shock = TRUE,
  iv_reg = FALSE, instrum = NULL, robust_se = FALSE,
  c_exog_data = NULL, l_exog_data = NULL, lags_exog_data = NaN,
  c_fd_exog_data = NULL, l_fd_exog_data = NULL,
  lags_fd_exog_data = NaN, confint = NULL, hor = NULL)
}
\arguments{
\item{data_set}{A \link{data.frame}, containing the entire panel data set. The first column has to be the
variable denoting the cross-section. The second column has to be the
variable denoting the time-section.}

\item{sample}{Boolean or character. Use full sample? TRUE (default). To estimate a subset, you have to provide
a sequence of those dates to use. This sequence has to be in the same format as the second column,
which contains the time-section.}

\item{endog_data}{Character. Name of the endogenous variable. You can only provide one endogenous variable at a time.}

\item{cumul_mult}{Boolean. Estimate cumulative multipliers? TRUE or FALSE. If TRUE, cumulative responses
are estimated via: \deqn{y_(t+h) - y_(t-1)} for h = 0,..., H-1.}

\item{shock}{Character. The column name of the variable to shock with.}

\item{diff_shock}{Boolean. Take first differences of the shock variable? TRUE or FALSE.}

\item{iv_reg}{Boolean. Use instrument variable approach? TRUE or FALSE.}

\item{instrum}{NULL or Character. The name(s) of the instrument variable(s) if iv_reg = TRUE.}

\item{robust_se}{Boolean. Estimate robust standard errors á la Driscoll and Kraay (1998)? TRUE or FALSE.}

\item{c_exog_data}{NULL or Character. Names of exogenous variables with contemporaneous impact.}

\item{l_exog_data}{NULL or Character. Names of exogenous variables with lagged impact.}

\item{lags_exog_data}{Integer. Lag length for the exogenous data with lagged impact.}

\item{c_fd_exog_data}{NULL or Character. Names of exogenous variables with contemporaneous impact of first differences.}

\item{l_fd_exog_data}{NULL or Character. Names of exogenous variables with lagged impact of first differences.}

\item{lags_fd_exog_data}{NaN or Integer. Number of lags for first differenced variables.}

\item{confint}{Double. Width of confidence bands. 68\% = 1; 90\% = 1.65; 9\% = 1.96.}

\item{hor}{Integer. Number of horizons for impulse responses.}
}
\value{
A list containing:

\item{irf_lin_mean}{A \link{matrix}, containing the impulse responses.
                  The columns are the horizons.}

\item{irf_lin_low}{A \link{matrix}, containing all lower confidence bands.
                  The columns are the horizons.}

\item{irf_lin_up}{A \link{matrix}, containing all upper confidence bands.
                                  The columns are the horizons.}

\item{reg_summaries}{Regression output for each horizon.}

\item{xy_data_sets}{Panel data sets with endogenous and exogenous variables for each horizon.}

\item{specs}{A list with data properties for the plot function.}
}
\description{
This function estimates impulse responses with local projections for panel data with fixed effects
             as in e.g. Jordà et al. (2018). It uses the \emph{plm()} function from the plm package to estimate the irfs.
              The function also allows to estimate cumulative impulse responses.
}
\examples{
\donttest{

# This example aims to replicate the STATA example, provided on
# Òscar Jordà's website (https://sites.google.com/site/oscarjorda/home/local-projections)
# It estimates the impulse reponse of the ratio of (mortgage lending/GDP) to a
# +1\% change in the short term interest rate

# Load libraries to download and read excel file from the website
 library(httr)
 library(readxl)
 library(dplyr)

# Retrieve the JST Macrohistory Database
 url_jst <-"http://www.macrohistory.net/JST/JSTdatasetR3.xlsx"
 GET(url_jst, write_disk(jst_link <- tempfile(fileext = ".xlsx")))
 jst_data <- read_excel(jst_link, 2L)

# Swap the first two columns so that 'country' is the first (cross section) and 'year' the
# second (time section) column
 jst_data <- jst_data \%>\%
             dplyr::select(country, year, everything())

# Prepare variables. This is based on the STATA 'data.do' file on Òscar Jordà's website
  data_set <- jst_data \%>\%
               mutate(stir     = stir)                         \%>\%
               mutate(mortgdp  = 100*(tmort/gdp))              \%>\%
               mutate(hpreal   = hpnom/cpi)                    \%>\%
               group_by(country)                               \%>\%
               mutate(hpreal   = hpreal/hpreal[year==1990][1]) \%>\%
               mutate(lhpreal  = log(hpreal))                  \%>\%

               mutate(lhpy     = lhpreal - log(rgdppc))        \%>\%
               mutate(lhpy     = lhpy - lhpy[year == 1990][1]) \%>\%
               mutate(lhpreal  = 100*lhpreal)                  \%>\%
               mutate(lhpy     = 100*lhpy)                     \%>\%
               ungroup()                                       \%>\%

               mutate(lrgdp    = 100*log(rgdppc))              \%>\%
               mutate(lcpi     = 100*log(cpi)) 		            \%>\%
               mutate(lriy     = 100*log(iy*rgdppc))           \%>\%
               mutate(cay      = 100*(ca/gdp))                 \%>\%
               mutate(tnmort   = tloans - tmort)               \%>\%
               mutate(nmortgdp = 100*(tnmort/gdp))             \%>\%

             # Prepare instrument
               mutate(lnarrow  = log(narrowm))                 \%>\%
               mutate(rlnarrow = lnarrow - lcpi)               \%>\%
               dplyr::select(country, year, mortgdp, stir, ltrate, lhpy,
                             lrgdp, lcpi, lriy, cay, nmortgdp, rlnarrow)


# Use sample from 1870 - 2013 BUT exclude WWI and WWII
  sample <-   seq(1870, 2016)[which(!(seq(1870, 2016) \%in\%
                              c(seq(1914, 1918),
                                seq(1939, 1947),
                                seq(2014, 2016))))]

# Estimate panel model
results_panel <-  lp_panel_lin(data_set          = data_set,
                               sample            = sample,
                               endog_data        = "mortgdp",
                               cumul_mult        = TRUE,

                               shock             = "stir",
                               diff_shock        = TRUE,
                               iv_reg            = FALSE,
                               instrum           = NULL,
                               robust_se         = TRUE,

                               c_exog_data       = "cay",
                               l_exog_data       = NULL,
                               lags_exog_data    = NULL,
                               c_fd_exog_data    = colnames(data_set)[c(seq(4,9),11)],
                               l_fd_exog_data    = colnames(data_set)[c(seq(3,9),11)],
                               lags_fd_exog_data = 2,

                               confint           = 1.67,
                               hor               = 10)

# Create and plot irfs
 plot_panel_lin <- plot_lin(results_panel)
 plot(plot_panel_lin[[1]])


# Create and add instrument to data_set
 set.seed(123)
 data_set   <- data_set \%>\%
               group_by(country) \%>\%
               mutate(instrument = 0.8*stir + rnorm(length(stir), 0, sd(na.omit(stir))/10)) \%>\%
               ungroup()


# Estimate panel model with iv
results_panel <-  lp_panel_lin(data_set          = data_set,
                               sample            = sample,
                               endog_data        = "mortgdp",
                               cumul_mult        = TRUE,

                               shock             = "stir",
                               diff_shock        = TRUE,
                               iv_reg            = TRUE,
                               instrum           = "instrument",
                               robust_se         = TRUE,

                               c_exog_data       = "cay",
                               l_exog_data       = NULL,
                               lags_exog_data    = NULL,
                               c_fd_exog_data    = colnames(data_set)[c(seq(4,9),11)],
                               l_fd_exog_data    = colnames(data_set)[c(seq(3,9),11)],
                               lags_fd_exog_data = 2,

                               confint           = 1.67,
                               hor               = 10)

# Create and plot irfs
 plot_panel_lin <- plot_lin(results_panel)
 plot(plot_panel_lin[[1]])

}

}
\references{
Driscoll, J.C., and Kraay, A.C. (1998). "Consistent Covariance Matrix Estimation with Spatially Dependent
Panel Data", \emph{Review of Economics and Statistics}, 80(4), pp. 549–560.

Jordà, Ò. (2005). "Estimation and Inference of Impulse Responses by Local Projections."
\emph{American Economic Review}, 95 (1): 161-182.

Jordà, Ò., Schualrick, M., Taylor, A.M. (2018). "Large and State-Dependent Effects of Quasi-Random Monetary Experiments",
\emph{NBER} working paper 23074, \emph{FRBSF} working paper 2017-02.
}
\author{
Philipp Adämmer
}
