% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/package.R
\docType{package}
\name{lpirfs-package}
\alias{lpirfs-package}
\title{Local Projection Impulse Response Functions}
\description{
lpirfs provides functions to estimate and plot linear as well as nonlinear impulse
responses based on local projections by Jordà (2005) <doi:10.1257/0002828053828518>.
Nonlinear impulse responses are estimated for two regimes which can be separated by a smooth transition function as
applied in Auerbach and Gorodnichenko (2012) <doi:10.1257/pol.4.2.1>, or by a simple dummy approach.
}
\author{
Philipp Adämmer
}
