% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{newey_west_c}
\alias{newey_west_c}
\title{Compute newey west covariance matrix}
\usage{
newey_west_c(y, x, h)
}
\arguments{
\item{y}{Numeric vector}

\item{x}{Numeric Matrix}

\item{h}{Integer}
}
\description{
Compute linear OLS coefficients and robust covariance matrix
based on Newey and West (1987).

Compute OLS parameters and robust standard errors based on Newey and West (1987).
The function is based on the Matlab code by James P. LeSage.
}
\references{
Newey W.K., and West K.D. (1987). “A Simple, Positive-Definite, Heteroskedasticity and
Autocorrelation Consistent Covariance Matrix.” \emph{Econometrica}, 55, 703–708.
}
