% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_lin_irfs.R
\name{plot_lin_irfs}
\alias{plot_lin_irfs}
\title{Compute and display plots of linear impulse responses}
\usage{
plot_lin_irfs(results_lin)
}
\arguments{
\item{results_lin}{A \link{list}() with 3D arrays estimated in \link{lp_lin}.}
}
\value{
A list with (gg-)plots for linear impulse responses.
}
\description{
Compute and display linear impulse responses estimated with \link{lp_lin}.
}
\examples{
\donttest{

# This function is deprecated. See plot_lin() instead.



 }
}
\author{
Philipp Adämmer
}
\keyword{internal}
