% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_lin_irfs.R
\name{plot_lin_irfs}
\alias{plot_lin_irfs}
\title{Compute (linear) irf plots}
\usage{
plot_lin_irfs(results_lin)
}
\arguments{
\item{results_lin}{A list with KXK plots constructed with ggplot2}
}
\description{
Compute (linear) irf plots estimated with \link{lin_lp}.
}
\examples{
 \dontrun{
# Load packages
  library(dplyr)
  library(doSNOW)
  library(parallel)
  library(Rcpp)

# Load data
  data_set_df <- interest_rules_var_data

# Create list for input
  specs <- list()

# Specify inputs
  specs$lags_lin       <- 4L
  specs$lags_criterion <- NaN
  specs$max_lags       <- 2L
  specs$trend          <- 0L
  specs$shock_type     <- 1L
  specs$confint        <- 1.96
  specs$hor            <- 12L

# Estimate model and save results
  results_lin  <- lp_lin(data_set_df, specs)

# Make and save plots
  linear_plots <- plot_lin_irfs(results_lin)

# Show single plots
  linear_plots[[1]]
  linear_plots[[2]]

# Show all plots
  library(ggpubr)
  library(gridExtra)

  lin_plots_all <- sapply(linear_plots, ggplotGrob)
  marrangeGrob(lin_plots, nrow = ncol(data_set_df), ncol = ncol(data_set_df))

 }


}
