% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_nl_irfs.R
\name{plot_nl_irfs}
\alias{plot_nl_irfs}
\title{Compute and display plots of nonlinear impulse responses}
\usage{
plot_nl_irfs(results_nl)
}
\arguments{
\item{results_nl}{A \link{list} with 3D arrays estimated in \link{lp_nl}.}
}
\value{
A list with (gg-)plots for nonlinear impulse responses.
}
\description{
Compute and display (nonlinear) impulse responses estimated with \link{lp_nl}.
}
\examples{
\donttest{
# Load package


 }
}
\author{
Philipp Adämmer
}
\keyword{internal}
