% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_nl_irfs.R
\name{plot_nl_irfs}
\alias{plot_nl_irfs}
\title{Function to plot irfs, estimated with lp_lin function}
\usage{
plot_nl_irfs(results_nl)
}
\arguments{
\item{results_nl}{List with estimated results from lp_nl function}
}
\value{
A list with (gg-)plots of impulse responses for two regimes
}
\description{
Function to plot irfs, estimated with lp_lin function
}
