% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/reduced_var.R
\name{reduced_var}
\alias{reduced_var}
\title{Function to estimate reduced VAR equation for equation and then shock matrix.}
\usage{
reduced_var(y_lin, x_lin, data_set_estim, specs)
}
\arguments{
\item{y_lin}{A list with endogenous variables.}

\item{x_lin}{A list with exogenous variables.}

\item{y_data}{A data frame for the exogenous variables.}

\item{lags}{Numeric value with number of lags.}
}
\value{
List with output from lm object.
}
\description{
Function to estimate reduced VAR equation for equation and then shock matrix.
}
