% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/reduced_var.R
\name{reduced_var}
\alias{reduced_var}
\title{Compute structural shock matrix via Cholesky decomposition}
\usage{
reduced_var(y_lin, x_lin, data_set_df, specs)
}
\arguments{
\item{y_lin}{A matrix with all endogenous variables}

\item{x_lin}{A matrix with lagged endogenous variables}

\item{specs}{A list with specifications from \link{lp_lin} or \link{lp_nl}}
}
\value{
Shock matrix (d)
}
\description{
Compute structural shock matrix via Cholesky decomposition with input variables
created in \link{lp_lin} or \link{lp_nl}.
}
\author{
Philipp Adämmer
}
