% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/switching_series.R
\name{switching_series}
\alias{switching_series}
\title{Function to estimate transition values}
\usage{
switching_series(switching_data, specs)
}
\arguments{
\item{switching_data}{A vector.}

\item{specs}{A list with specifications for \link{lp_nl}().}
}
\value{
\item{fz:}{A vector with values from smooth transition function for \eqn{F(z_{t-1})}.}
}
\description{
Function to estimate transition values. It uses a smooth transition function as
proposed by Auerbach & Gorodnichenko (2012). The time series, which is used in the transition function,
can be de-trended via the Hodrick-Prescott (see Auerbach and Gorodnichenko, 2013).
}
\references{
Auerbach, A. J., and  Gorodnichenko Y. (2012). "Measuring the Output Responses to Fiscal Policy."
\emph{American Economic Journal: Economic Policy}, 4 (2): 1-27.

Auerbach, A. J., and Gorodnichenko Y. (2013). "Fiscal Multipliers in Recession and Expansion."
\emph{NBER Working Paper Series}. Nr 17447.
}
\author{
Philipp Adämmer
}
